% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentumqbs181.R
\name{momentumQBS181}
\alias{momentumQBS181}
\title{A momentum function}
\usage{
momentumQBS181(ticker, from_, to_)
}
\arguments{
\item{from_}{the starting time of the tickers}

\item{to_}{the ending time of the tickers}

\item{tickers}{The names of tickers you want to read from yahoo.(like:'AAPL,GOOGL,NFLX,SQ,TSLA')}
}
\description{
This function allows you to calculate the stocks returns based on 30/60/120
days momentum strategies, and will plot the return curves.
You need to library('SIT') first.
}
\author{
Zhang, Shao, Wang, Liang
}
